[BEAM-9682] Move existing Go SDK katas fixed time window lesson into windowing section (#13008)
