+++ learning/katas/go/core_transforms/section-info.yaml
@@ -28,4 +28,3 @@ content:
 - additional_outputs
 - branching
 - composite
-- windowing
