--- learning/katas/go/core_transforms/windowing/lesson-remote-info.yaml
@@ -1,3 +1,0 @@
-id: 387853
-update_date: Thu, 06 Aug 2020 17:53:20 UTC
-unit: 377026
