--- learning/katas/go/core_transforms/windowing/windowing/task-remote-info.yaml
@@ -1,2 +1,0 @@
-id: 1464828
-update_date: Thu, 27 Aug 2020 00:08:00 UTC
