+++ learning/katas/go/course-remote-info.yaml
@@ -1,2 +1,2 @@
 id: 70387
-update_date: Wed, 30 Sep 2020 15:51:15 UTC
+update_date: Wed, 30 Sep 2020 15:56:05 UTC
