+++ learning/katas/go/windowing/fixed_time_window/fixed_time_window/task-info.yaml
