+++ learning/katas/go/windowing/fixed_time_window/fixed_time_window/task-remote-info.yaml
@@ -1,0 +1,2 @@
+id: 1605362
+update_date: Sat, 03 Oct 2020 19:31:13 UTC
