+++ learning/katas/go/windowing/fixed_time_window/lesson-remote-info.yaml
@@ -1,0 +1,3 @@
+id: 414186
+update_date: Sat, 03 Oct 2020 19:31:08 UTC
+unit: 403690
