+++ learning/katas/go/windowing/section-info.yaml
@@ -18,4 +18,4 @@
 #
 
 content:
-- windowing
+- fixed_time_window
