+++ learning/katas/go/windowing/section-remote-info.yaml
@@ -1,0 +1,2 @@
+id: 150684
+update_date: Sat, 03 Oct 2020 19:31:07 UTC
